+++ src/app/dashboard/mou/page.tsx
@@ -9,7 +9,7 @@ import {
   Button,
 } from "@mui/material";
 import { Dialog } from "@mui/material";
-import Pagination from "@/components/ui/Pagination";
+import Pagination from "@/components/ui/Navigation/Pagination";
 import PermissionGuard from "@/components/PermissionGuard";
 import MouList from "@/components/mou/MouList";
 import { useMou } from "@/hooks/useMou";
