+++ src/components/leads/CheckUploadStatusDialog.tsx
@@ -19,7 +19,7 @@ import {
 import CloseIcon from "@mui/icons-material/Close";
 import RefreshIcon from "@mui/icons-material/Refresh";
 import DownloadIcon from "@mui/icons-material/Download";
-import Pagination from "../ui/Pagination"; // Pagination component
+import Pagination from "../ui/Navigation/Pagination"; // Pagination component
 
 interface Upload {
   _id: string;
