+++ src/components/ui/avatar/index.ts
@@ -1,0 +1,1 @@
+export {};
