+++ src/components/ui/dialogs/index.ts
@@ -1,0 +1,1 @@
+export {};
