+++ src/components/ui/profile/index.ts
@@ -1,0 +1,1 @@
+export {};
