+++ src/components/ui/provider/MuiRootProvider/index.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import React, { ReactNode } from "react";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+
+interface MuiRootProviderProps {
+  children: ReactNode;
+}
+
+const theme = createTheme({
+  palette: {
+    mode: "light",
+    primary: {
+      main: "#1976d2",
+    },
+    secondary: {
+      main: "#9c27b0",
+    },
+  },
+});
+
+const MuiRootProvider: React.FC<MuiRootProviderProps> = ({ children }) => {
+  return (
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      {children}
+    </ThemeProvider>
+  );
+};
+
+export default MuiRootProvider;
