+++ src/components/ui/table/index.ts
@@ -1,0 +1,1 @@
+export {};
